requirements.txt
Django==5.1
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
drf-spectacular==0.27.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
pep8-naming==0.13.2
requests==2.32.3
